backend/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
httpx==0.27.2
openai==1.43.0
python-dotenv==1.0.1
pydantic==2.9.2
supabase==2.6.0
pytest==8.3.3
pytest-asyncio==0.24.0
anyio==4.6.2.post1
tenacity==9.0.0
orjson==3.10.7
itsdangerous==2.2.0
starlette==0.38.5
